Fix some test definitions after changes in extension parsing
ScExtensions no longer have ScTemplateBody in them

+++ scala/scala-impl/src/org/jetbrains/plugins/scala/lang/parser/parsing/expressions/Expr1.scala
@@ -342,6 +342,8 @@ object Expr1 extends ParsingRule {
     val iw = builder.currentIndentationWidth
     builder.advanceLexer() //Ate if
     builder.getTokenType match {
+      case InScala3(_) if parseParenlessIfCondition() =>
+      // already parsed everything till after 'then'
       case ScalaTokenTypes.tLPARENTHESIS =>
         builder.advanceLexer() //Ate (
         builder.disableNewlines()
@@ -349,13 +351,10 @@ object Expr1 extends ParsingRule {
         builder.getTokenType match {
           case ScalaTokenTypes.tRPARENTHESIS =>
             builder.advanceLexer() //Ate )
-            if (builder.getTokenType == ScalaTokenType.ThenKeyword) builder.advanceLexer()
           case _ =>
             builder error ErrMsg("rparenthesis.expected")
         }
         builder.restoreNewlinesState()
-      case InScala3(_) if parseParenlessIfCondition() =>
-      // already parsed everything till after 'then'
       case _ =>
         builder error ErrMsg("condition.expected")
     }
@@ -384,18 +383,24 @@ object Expr1 extends ParsingRule {
   }
 
   def parseParenlessIfCondition()(implicit builder: ScalaPsiBuilder): Boolean = {
-    val rollbackMarker  = builder.mark()
-    val cond = ExprInIndentationRegion()
+    val startedWithLParen = builder.getTokenType == ScalaTokenTypes.tLPARENTHESIS
+    val rollbackMarker = builder.mark()
+    val success = ExprInIndentationRegion() && (
+      if (builder.getTokenType == ScalaTokenType.ThenKeyword) {
+        builder.advanceLexer()
+        true
+      } else if (startedWithLParen) {
+        false
+      } else {
+        builder error ErrMsg("expected.then")
+        true
+      }
+      )
 
-    if (!cond) {
-      rollbackMarker.rollbackTo()
-      false
-    } else {
-      if (builder.getTokenType != ScalaTokenType.ThenKeyword) builder.error(ScalaBundle.message("expected.then"))
-      else                                                    builder.advanceLexer()
-      rollbackMarker.drop()
-      true
-    }
+    if (success) rollbackMarker.drop()
+    else rollbackMarker.rollbackTo()
+
+    success
   }
 
   def parseMatch(exprMarker: PsiBuilder.Marker)(implicit builder: ScalaPsiBuilder): Unit = {
